Also locate and obey modelines at the end of the file.
* vimvars.el (vimvars-check-lines): We also check
vimvars-check-lines lines at the bottom of the file.
(vimvars-obey-this-vim-modeline): Factor out of
vimvars-obey-vim-modeline in order to separate the concerns of
finding the modeline and obeying it.
(vimvars-obey-top-modeline): New function, locates a modeline at
the top of the file.
(vimvars-obey-bottom-modeline): New funciton, locates a modeline
at the bottom of the file.
(vimvars-obey-vim-modeline): Move some of the body into
vimvars-obey-this-vim-modeline.  Call vimvars-obey-top-modeline
and (if it fails to find a modeline) vimvars-obey-bottom-modeline.
(vimvars-enable-feature): Fix typo in docstring.